+++ src/code/launcher/gui_launcher.cpp
@@ -11,166 +11,11 @@
 #include "../lang.h"
 #include "pcsx_interceptor.h"
 #include "gui_btn_guide.h"
+#include "../gui/renderText.h"
+#include "NotificationLine.h"
 #include <algorithm>
 
-void PsCarouselGame::loadTex(SDL_Renderer *renderer) {
-    shared_ptr<Gui> gui(Gui::getInstance());
-
-    if (coverPng == nullptr) {
-        SDL_Texture *renderSurface = SDL_CreateTexture(renderer, SDL_PIXELFORMAT_RGBA32, SDL_TEXTUREACCESS_TARGET, 226,
-                                                       226);
-        SDL_SetRenderTarget(renderer, renderSurface);
-        SDL_RenderClear(renderer);
-        SDL_Rect fullRect;
-
-        string imagePath = (*this)->folder + Util::separator() + (*this)->base + ".png";
-        if (Util::exists(imagePath)) {
-            coverPng = IMG_LoadTexture(renderer, imagePath.c_str());
-        } else coverPng = nullptr;
-
-        if (coverPng != nullptr) {
-            fullRect.x = 0;
-            fullRect.y = 0;
-            fullRect.h = 226, fullRect.w = 226;
-
-            Uint32 format;
-            int access;
-            SDL_QueryTexture(coverPng, &format, &access, &fullRect.w, &fullRect.h);
-
-            SDL_Rect outputRect;
-            if (gui->cdJewel != nullptr) {
-                outputRect.x = 23;
-                outputRect.y = 5;
-                outputRect.h = 217;
-                outputRect.w = 199;
-            } else {
-                outputRect.x = 0;
-                outputRect.y = 0;
-                outputRect.h = 226;
-                outputRect.w = 226;
-            }
-            if (coverPng != nullptr) {
-                SDL_RenderCopy(renderer, coverPng, &fullRect, &outputRect);
-            }
-            if (coverPng != nullptr) {
-                SDL_DestroyTexture(coverPng);
-                coverPng = nullptr;
-            }
-
-            fullRect.x = 0;
-            fullRect.y = 0;
-            fullRect.h = 226, fullRect.w = 226;
-            if (gui->cdJewel != nullptr) {
-                SDL_RenderCopy(renderer, gui->cdJewel, &fullRect, &fullRect);
-            }
-            coverPng = renderSurface;
-        }
-        SDL_SetRenderTarget(renderer, nullptr);
-    }
-}
-
-void PsCarouselGame::freeTex() {
-    if (coverPng != nullptr) {
-        SDL_DestroyTexture(coverPng);
-        coverPng = nullptr;
-    }
-}
-
-
-// global renderText
-void renderText(SDL_Renderer * renderer, int x, int y, const std::string & text, const SDL_Color & textColor, TTF_Font *font, bool center, bool background) {
-    int text_width;
-    int text_height;
-    SDL_Surface *surface;
-    SDL_Texture *texture;
-    SDL_Rect rect;
-
-    if (text.size() == 0) {
-        texture = SDL_CreateTexture(renderer, SDL_PIXELFORMAT_RGBA32, SDL_TEXTUREACCESS_STATIC, 0, 0);
-        rect.x = 0;
-        rect.y = 0;
-        rect.h = 0;
-        rect.w = 0;
-    } else {
-        surface = TTF_RenderUTF8_Blended(font, text.c_str(), textColor);
-        texture = SDL_CreateTextureFromSurface(renderer, surface);
-        text_width = surface->w;
-        text_height = surface->h;
-        SDL_FreeSurface(surface);
-        rect.x = x;
-        rect.y = y;
-        rect.w = text_width;
-        rect.h = text_height;
-    }
-    SDL_Rect inputRect;
-    inputRect.x = 0;
-    inputRect.y = 0;
-    inputRect.w = rect.w;
-    inputRect.h = rect.h;
-
-    if (center) {
-        rect.x = 640 - (rect.w / 2);
-    }
-
-    if (background) {
-        SDL_SetRenderDrawColor(renderer, 0, 0, 0, 70);
-        SDL_Rect backRect;
-        backRect.x = rect.x - 10;
-        backRect.y = rect.y - 2;
-        backRect.w = rect.w + 20;
-        backRect.h = rect.h + 4;
-
-        SDL_RenderFillRect(renderer, &backRect);
-    }
-
-    SDL_RenderCopy(renderer, texture, &inputRect, &rect);
-    SDL_DestroyTexture(texture);
-};
-
-void NotificationLine::setText(string _text, bool _timed, long _timeLimit, const SDL_Color & _textColor, TTF_Font *_font) {
-    text = _text;
-    timed = _timed;
-    notificationTime = SDL_GetTicks();  // tick count when setText called
-    if (notificationTime == 0)  // if by chance it's 0.  0 flags that the timeLimit has been reached and to turn off the display
-        ++notificationTime;
-    timeLimit = _timeLimit;
-    textColor = _textColor;
-    font = _font;
-};
-
-void NotificationLine::setText(string _text, bool _timed, long _timeLimit) {
-    setText(_text, _timed, _timeLimit, textColor, font);
-};
-
-void NotificationLine::tickTock(SDL_Renderer * renderer) {
-    if (timed) {
-        if (notificationTime != 0) {
-            long currentTimeTicks = SDL_GetTicks();
-            if (currentTimeTicks - notificationTime > timeLimit) // if time limit reached
-                notificationTime = 0;   // turn off the display
-        }
-        if (notificationTime != 0)
-        ::renderText(renderer, x, y, text, textColor, font, true, true);
-    } else // not timed - keep display on
-        ::renderText(renderer, x, y, text, textColor, font, true, true);
-}
-
-void NotificationLines::createAndSetDefaults(int count, int x_start, int y_start, TTF_Font * font, int fontHeight, int separationBetweenLines) {
-    for (int line=0; line < count; ++line) {
-        NotificationLine notificationLine;
-        notificationLine.font = font;
-        notificationLine.textColor = brightWhite;
-        notificationLine.x = x_start;
-        notificationLine.y = y_start + (line * (fontHeight + separationBetweenLines));
-        if (line == 0)
-            notificationLine.timed = false; // keep line 0 displayed
-        else {
-            notificationLine.timed = true;
-            notificationLine.timeLimit = DefaultShowingTimeout;
-        }
-        lines.push_back(notificationLine);
-    }
-}
+using namespace std;
 
 bool wayToSort(const PsGamePtr &i, const PsGamePtr &j) { return SortByCaseInsensitive(i->title, j->title); }
