+++ src/code/launcher/ps_game.h
@@ -8,8 +8,8 @@
 #include <SDL2/SDL_image.h>
 #include <SDL2/SDL_mixer.h>
 #include <SDL2/SDL_ttf.h>
-#include "ps_carousel.h"
 #include <string>
+#include <vector>
 #include <memory>
 
 class PsGame {
